python-worker/requirements.txt
python-pptx==1.0.2
pytesseract==0.3.13
Pillow==11.2.1
